Fix ambiguous URIs in function server API
Airlift's HttpUriBuilder expects that uris passed to it are
not yet encoded. However, because we use a semicolon ';' to
delimit function names and parameters in the function server
the semicolon which is encoded to %3B gets double-encoded to
%253B by the URI builder (% encodes to %25). Older versions
of Jetty allow this sequence in the URI. Newer servers however
reject this request encoded with %25xxx with an
error stating "AMBIGUOUS_PATH_PARAMETER".

If we avoid using the HttpUriBuilder, we don't double-encode
the semicolon, and the requests will work against the new
server