Fix ambiguous URIs in function server API
In Java 17, when URI.create is called on a a String, the string
is URL encoded. This will cause encoded strings to be encoded
twice if the input is already URL encoded. For example %3B will
become %253B to encode the '%' character.

Newer Jetty servers reject requests encoded with %25xxx
as "AMBIGUOUS_PATH_PARAMETER".

However, if the String is not URL encoded, it will be encoded
as normal in URI.create. The ; character is actually allowed as a
delimiter  for path parameters. See RFC 3986 S3.3 which means
that when passing it directly to URI.create, the semicolon is
left in place and not encoded.

This presents a problem for the  function server API since we will
not be able to create URIs easily that are encoded. This change
updates the delimiter used to be `|` which is not in the reserved
URI delimiter list.