feat: Pass mapping from PlanNodeId to ExchangeClient to TaskListener::onCompletion (#13131)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13131

We're using TaskListener to log query plan fragments so that we can rerun the query
later. To do that, we need to know the relation among plan fragments, i.e., the
producers and consumers of shuffles. This diff makes TaskListener::onCompletion
receives a map from consumer PlanNode id to a ExchangeClient with which we can
log the mapping from consumer PlanNode id to a list of producer Task id in
TaskListener.

Reviewed By: kevinwilfong

Differential Revision: D69828745

fbshipit-source-id: 70be537a71b310579ed38cab458626f7f11edba0