fix: Ensure constant filter results are properly handled in processFilterResults (#13155)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13155

In tableScan with filtering, we can get an edge case in `HiveDataSource::next()` such that when we call `processFilterResults` on a `filterResult` with constant encoding, we may not be setting the `FilterEvalCtx.selectedIndices` correctly. As a consequence, when we wrap the vector in `HiveDataSource`, we either crash, get undefined behavior, or fail at dictionary construction time.

The fix here is that if the `selectivityVector` did not have all of its rows selected, we will update the indices again. This ensures that the selectedIndices and rows are correctly updated. Essentially, this addresses the edge case where `numSelected = filterResult->size()`.

Reviewed By: Yuhta

Differential Revision: D73636993

fbshipit-source-id: 93b5b66c9cfdc2d8649c261e9f1c0d1422935afc