refactor(StringView): Remove unneeded compare condition (#13173)
Summary:
`isInline() && other.isInline()` can already ensure that both operands
are inline, and because `StringView::isInline()` is an inline function,
it is also efficient enough, so the condition `size <= kInlineSize` is
not needed. This PR removes the useless condition.

Also, since `size` is not actually the smaller of the two lengths (note
that it is minus `kPrefixSize` above), the meaning of this condition
is unclear.

No functional changes.

Fixes https://github.com/facebookincubator/velox/issues/13172

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13173

Reviewed By: Yuhta

Differential Revision: D73794861

Pulled By: kevinwilfong

fbshipit-source-id: 781913d52a11e6c6adee583c7832e6245fd34fab