feat: Add option to produce filters as expressions (#13178)
Summary:
Adds an option to TableScanBuilder and TpchQueryBuilder to produce filters that would be pushed down into a scan as a filter node on top of the scan instead.  This is used in testing Verax so that the optimizer sees the filters as discrete and can add new filters to these.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13178

Reviewed By: Yuhta

Differential Revision: D73809820

Pulled By: oerling

fbshipit-source-id: 44a8f64bb9f2865d59e5268deec57166b77a6aba