Separate native IndexJoinRewriter from Java path
Index Join has a slightly different representation in the Velox query
plan compared to Presto Java, as well as its execution model. It is
hard to make changes to the optimizer without affecting the Java path.
Therefore, this change breaks the IndexJoinRewriter into two to
decouple native query planning for index join from Java.

Currently, there is one main difference:

- `IndexJoinNode` doesn't support in-line filtering on Java execution,
so if the original `Join` node has a filter, it is needed to add a
`FilterNode` to the query plan, while it is not necessary for Velox as
the filter can be expressed as `joinConditions` in Velox
`IndexLookupJoinNode`.

- In addition, if the join type is `LEFT`/`RIGHT` while a filter is
present, the join is ineligible for Index Join optimization for Java,
while it is still eligible for native execution.