fix: Fix the child vector resize length in row vector (#13195)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13195

Row vector resize first clears the child vector sizes to zero by preparing for reuse. When resize to new length for the child vector, it checks the row vector's size which leave the child vector size as zero if the row vector's new size <= the old size. This PR fixes this by setting the child vector size correctly.

Reviewed By: yuandagits

Differential Revision: D73695918

fbshipit-source-id: 3d199bb60998e7f5d7e04031e7835ee5e244d1f6