misc: Add support for TimestampWithTimeZone in Aggregation, Window, and Join Fuzzers (#11897)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11897

This change adds support for TimestampWithTimeZone in AggregationFuzzer as partition keys and aggregates (it cannot be a
sorting key as it does not support +/-) with PrestoQueryRunner.

I added a function inputProjections to ReferenceQueryRunner which takes in the input batch and uses the IntermediateTypeTransform introduced earlier to do a
conversion and generate the expressions.

The rest of this change is just plumbing these projections into all plans throughout the fuzzers, as well
as the plans generated in the ResultVerifiers.

Reviewed By: kagamiori

Differential Revision: D67346942

fbshipit-source-id: 100903795347255d03ff878cf956033a06b7ff81