refactor: Refactor the logic for constructing Join plans from JoinFuzzer into JoinMaker (#13046)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13046

JoinFuzzer has become monolithic (>1500 lines) and functional. We have a lot of functions to
construct join plans that take the same or very similar long lists of arguments. These functions
also do a lot of the same things but do not share code. It makes it very difficult to answer
basic questions like: what join plans are we testing in JoinFuzzer?

To try to improve this I've added a JoinMaker class. This class holds the state that's needed to
construct the various plan alternatives, e.g. the inputs, the join type, the filter, and the output
columns. This way we can call clear functions like makeHashJoin and the arguments specify
just what adjustments we're making to the join plan (are we using flat inputs, are we reversing
the order of the inputs, are we partitioning the inputs, etc.)

In addition I saw we were passing a lot of the same arguments for the probe and build sides of
the join. So I added a JoinSource class to capture that (the input data, the key columns, the
type of the data, etc.). Besides making it cleaner I think this will help support fuzzing plans
that join more than 2 tables eventually.

Reviewed By: kagamiori

Differential Revision: D73044135

fbshipit-source-id: 89febad9019a93543605d6abfc951b95c89481dc