fix: Prevent copying vector of shared_pointer velox vectors (#13200)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13200

Looking at round profiles, saw ~20% cycles on allocating and de-allocing vector of shared_ptr<velox vectors> on container serde path. Likely happens when we have many many vectors in the child and this function is called a lot.

Internal profile shows 10% cycles spent when sampled on 5 hosts in a fleet of 300.

Reviewed By: Yuhta

Differential Revision: D73932727

fbshipit-source-id: c9bb3e33623c7573cf94fe8c23fa9d1702ac864b