fix: Change pushback global arbitration to be based on promised capacity (#13231)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13231

Global arbitration (both shrink capacity API and bg global arbitration) make decisions on sufficient arbitrated capacity by immediately freed capacity. This way it is often the arbitrator over commit targets because the soon to be aborted query capacity is not accounted, resulting in more queries got killed. Change it to take into consideration of the part soon to be freed by query abortion as well.

Reviewed By: xiaoxmeng

Differential Revision: D74186831

fbshipit-source-id: 854664bbc454fc46ffd1fb5b4a299e0547674bc0