fix: AggregationFuzzer fails when replaying plans with intermediate type transforms (#13259)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13259

When replaying a plan persisted after a failure, it assumes that the AggregationNode is
immediately preceded by the ValuesNode.  With intermediate type transforms we insert a
projection between these nodes causing it to fail.

To fix this I updated the code to allow any number of nodes between the AggregationNode
and the ValuesNode, as long as they only have a single source (I think this code would need
to be rewritten to handle a case with more than one sources anyway).

Reviewed By: kgpai

Differential Revision: D74340284

fbshipit-source-id: 4643ce75039da33563282c3b252caa48c789d26d