refactor: TopNRowNumber::getOutputFromMemory loop (#11440)
Summary:
The getOutputFromMemory loop first output the remainder of rows (if any) from the current partition. If there was still space on the output buffer, then it went into a loop trying to add as many partitions (some with all rows in entirety and a last partial one possibly). The code is simplified to just a loop that carries forward from the current partition to fill output rows. The loop iterates over as many partitions can be output like the one before.

The priority queue in the TopNRowNumber partitions pops rows in reverse order of row number.  The old code maintained a remainingRowsInPartition_ variable that was used in some odd ways to compute the 'start' rowNumber for each output block from a partition. This is not needed. The partition.rows.size() can be used simply to get the correct row numbers.

Remove the un-necessary currentPartition() function as a result.

ref https://github.com/facebookincubator/velox/pull/11554 for TopNRank enhancements to this operator.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11440

Reviewed By: xiaoxmeng

Differential Revision: D74427134

Pulled By: kagamiori

fbshipit-source-id: d4e422a899db733565846e844834dd6bb408b689