feat: Zip the sorting column indices and compare flags (#13269)
Summary:
Currently, the sorted spiller moves the sorting columns to the front
and uses the number of sorting keys to access them. Zip the sorting
column indices and compare flags into a vector of pairs, decoupling
the RowVector layout from the sorting process. Hence, it can handle
both cases: when the sorting keys are adjusted and when no adjustment
is made.

Part of https://github.com/facebookincubator/velox/issues/13260

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13269

Reviewed By: tanjialiang

Differential Revision: D74449348

Pulled By: xiaoxmeng

fbshipit-source-id: 44ecfc266168bdd4c55b8bd2d9b5df1f01a328e0