Fix protocol generation to properly remove license
The java-to-struct-json.py script was supposed to remove the license
header from special classes ending in .inc so they can be inserted in
protocol definition sources. It was incorrectly using the Velox license
header, which is slightly different, and never removed the license.
This uses the correct license.header file from presto_cpp so it is
removed from special classes when generating the presto protocol.