feat: Support buffering in local exchange operator (#13234)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13234

When the number of output partitions is low, it is preferred to process
one input vector at a time. For example, with 10 output partitions
splitting a single 100KB input vector into 10 10KB vectors is acceptable.
However, when the number of output partitions is high it may result in a
large number of tiny vectors generated. For example, with 100 output
partitions splitting a single 100KB input vector results in 100 1KB
vectors. Exchanging and processing tiny vectors may negatively impact
performance. To avoid this, buffered partitioning is used to accumulate
larger vectors.

Reviewed By: Yuhta

Differential Revision: D74205073

fbshipit-source-id: 069d5d94374b554c205905b447f57055bbddaa4a