fix: FloatingPointRange::testDoubleRange (#13316)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13316

When either side of the filter range is unbounded, and the
corresponding upper/lower is set to a finite value,
`FloatingPointRange::testDoubleRange` would use the upper/lower value and ignore
the fact that it's unbounded on that side.  In Prestissimo the upper/lower is
set to infinity or negative infinity if the side is unbounded, so we do not see
this bug there.

Reviewed By: xiaoxmeng

Differential Revision: D74538066

fbshipit-source-id: 3a60ecada185f13c6878fd23777fbcf2b60c9f63