fix: Support float to double schema evolution (#13317)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13317

Currently the Nimble selective reader is missing the schema evolution support for `float` to `double` conversion (other types are already supported).  Adding this missing support using the existing mechanism provided by the selective reader framework.

There is a caveat that when this schema evolution happens, the current `Filter::testValues` implementation would not work (which means we cannot use fast path).  In DWRF we have a dedicated `FloatingPointDecoder` that knows the requested and data type, but in Nimble, decoders are decoupled from column reader, so we cannot pass that information to decoders.  Instead we add `hasBulkPath` as a compile time constant to the decoder visitor, so it can disable the fast path when needed.
bypass-github-export-checks

Reviewed By: HuamengJiang

Differential Revision: D74537747

fbshipit-source-id: 3d949cba3c55884297017cf5455dadeb6b264c04