fix: Optimize array and map copy in encodedVectorCopy (#13320)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13320

In `toNestedRanges` and `compactNestedRanges`, we were sorting the
ranges just in case the ranges are not ordered and we missed the opportunity to
merge adjacent ranges together.  But it turns out the sorting itself is
expensive and in most of the case the ranges are already sorted, so we skip the
sorting phase and directly iterate on the ranges and merge only the existing
adjacent ones.

Reviewed By: bikramSingh91

Differential Revision: D74582120

fbshipit-source-id: 3df55eab04d9f381cb726961ff0f5eb327396dc9