test: Add support for TimestampWithTimeZone in Expression Fuzzer (#12805)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12805

This change adds support for TimestampWithTimeZone in ExpressionFuzzer with
PrestoQueryRunner.

This change primarily just integrations the inputProjections function from
ReferenceQueryRunner into the ExpressionFuzzer and plumbs the resulting updated input
Vectors and projections through to verification.

The only other change was to VectorFuzzer, this exposed some code paths that were not
respecting the input generators for custom types.

Reviewed By: kagamiori

Differential Revision: D71849125

fbshipit-source-id: 80fa6e573043eedebe76a56ec71b9f6e9b5f402f