fix(parquet): Avoid SEGV if table column type does not match file column type (#12350)
Summary:
If a user defines a table, for example in Hive, where the column types don’t match the file column types SEGV might occur. Specifically, the SEGV has been observed if the parquet file contains a VARCHAR column but the table defines an INTEGER column instead and the data is accessed via TableScan using a basic select * .

The resulting vector is of type string but it doesn’t match the table metadata and can cause a SEGV in the PartitionedOutput operator.

This also prevents issues and error coming from the readers when they encounter types that are not part of the switch, for example, defining a VARCHAR type column when the file column is an INTEGER.

Fixes: https://github.com/facebookincubator/velox/issues/12349

Pull Request resolved: https://github.com/facebookincubator/velox/pull/12350

Reviewed By: kKPulla

Differential Revision: D69786341

Pulled By: pedroerp

fbshipit-source-id: 2de2fcb90b9657e3e758af656a891a819ec8d0ac