fix: Access after buffer boundary causing crash in selective reader (#13344)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13344

In case we are filtering `int16_t` data, we read 16 of them at same
time.  However the row numbers are 32 bits long and we can only do them 8 at a
time.  In case that all 16 rows passing the filter, we read and write 2 SIMD
batches of indices without checking the end boundary.  This is ok for values (we
always ensure 1 SIMD batch padding) but not ok for indices (need 2 batches
padding in worst case), causing accessing over boundary and crashes.  This was
not generating incorrect results because the length of the output rows is set
correctly and extra rows at end are ignored.
bypass-github-export-checks

Reviewed By: amitkdutta, HuamengJiang

Differential Revision: D74752697

fbshipit-source-id: 110b6efbfe9ddf6c7afbc921fea6e99bb96d66ef