feat: Clean up error reporting for Geometry functions (#13352)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13352

This cleans up the error handling for the Geometry serde functions,
to establish the pattern for later geometry functions.  In this, we catch
all exceptions from GEOS and raise them as a UserError, so users can bypass
them with TRY as desired.

Certain GEOS exceptions are returned as internal errors, like
AssertionFailedException and UnsupportedOperationException, because they
indicate bugs in either Geos code or our code.

We also let errors from the internal de/serialization bubble up, since those
are bugs in our code.

Reviewed By: bikramSingh91

Differential Revision: D74761810

fbshipit-source-id: 1570a7714cc0a5e2fe40a0f8a4a99addbd57ae5f