fix: Loose mixed grouped mode join condition (#13355)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13355

Today we tell if a join is in mixed grouped mode by telling if the entire task is mixed grouped mode. Telling based on the task is a stricter condition than the join node itself because a task can be mixed mode but the a join node inside the task can be non-mixed mode. This disallows some supported queries to run. This change looses the condition by replacing the task scope mixed mode method with a more specific method that tells the mode of a particular join node.

Reviewed By: xiaoxmeng

Differential Revision: D74839716

fbshipit-source-id: 1314b5af693e998c965f98fb6a21995a64009d50