Fix ClusterManager failing to poll
Previously, ClusterManager relies on CompletableFuture.supplyAsync
which uses the ForkJoin commonPool to execute tasks. On small
machines this pool has a limited number of threads and the previous
implementation used an infinite loop that would never break. This
caused a starvation of resources which could cause tests to hang and
due to the check never completing.

The new implementation uses a single-threaded scheduled executor to
prevent starvation and to properly setup/teardown resources. It also
improves the test flakiness by ensuring the file watch service starts
picking up changes to the config files first before moving on to
later test cases.