Always build the docs
There was a build failure on trunk because the docs failed to build.  Docs only build currently if there's a change detected in the docs directory.  However, there are some subtle interactions between the docs and the libraries and code that we use, so it's safest to always build the docs to prevent issues such as #25177.