fix(json): Bug with duplicate keys leading to incorrect writes (#13445)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13445

Currently when we cast json to row's if we encounter a duplicate key, we throw, however this throw isnt caught in the json cast row calling function , instead the throw is subsumed by the applyToselectedNoThrow call. This causes a problem because the previous write operation is not cancelled and is still valid, often leading to invalid data being read. This diff fix this bug. This bug is very evident when the key duplicated has a string value  , it might be harder to replicate if the key duplicated is against a complex type , however the fix should work for all types.

Reviewed By: kevinwilfong

Differential Revision: D75251964

fbshipit-source-id: ae7889cad7dd90f82ce44eb3814485e190700b45