feat: Support split barrier processing for Veloski (#13434)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13434

To support barrier processing in Koski, we introduce a BarrieredSplitProvider that takes a split iterator on construction. It provides an API called startNextBarrier(), which retrieves the next split from the split provider and deserializes it into a map with one entry per data source. The startNextBarrier() method is invoked when the Velox task begins a new barrier processing cycle.
The sequential Veloski task execution process is as follows:

1. Add Koski splits to all Velox task leaf nodes.
2. Obtains the barrier future from the Velox task.
3. Run a loop to retrieve output batches from the Velox task. When a future is returned from the Velox task Next() call, check if the barrier future is ready. If not, wait directly on the Next() output future. Otherwise, it indicates that the Velox task has completed a barrier processing cycle and calls startNextBarrier() from the split provider. This method returns a boolean indicating whether a split was fetched from the split iterator (true) or not (false). If true, add the next Koski splits to all leaf source nodes and wait on the output batch future to continue data processing. If false, send a no more split signal to terminate Velox task execution.

For the Koski connector, we may want to reuse the DWIO batch scan operator created within the Koski data source across multiple splits. We only need to reset the operator if the previous split was not fully processed (i.e., the DWIO batch scan drained flag is not set). This scenario occurs when we drop an operator input, as mentioned earlier for the MergeJoin case.

Reviewed By: littlesamo, Yuhta

Differential Revision: D73720670

fbshipit-source-id: 7a7500f9365d69ea4b3c3087369a5ec96be04b83