test(type): Add a unittest for NoisyCountIfGaussianAggregationTest (#13443)
Summary:
### Diff Summary

#### Code Change in NoisyCountIfGaussianAggregationTest.cpp
The code change adds a new unit test to `NoisyCountIfGaussianAggregationTest.cpp` for the case when there are zero rows in the input, the aggregation is performed on multiple groups, and a constant noise scale is used. This test case aims to ensure the robustness of the aggregation function in handling edge cases with no input data while still providing outputs for different groups. This test is essential in providing insights into how the function handles these specific conditions and helps in maintaining the overall reliability of the aggregation function in a variety of scenarios.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13443

Reviewed By: amitkdutta

Differential Revision: D75242440

fbshipit-source-id: a69508e8594877967d0408653d493a394cd0393c