fix: Remove incorrect check in ArbitrationParticipant::abortLocked() (#13513)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13513

The check is incorrect because it could be possible that a driver thread and global arbitration thread is aborting at the same time. Driver thread aborts through reclaim exception, and could be faster and could already have set the aborted_ to true, making the check on global arbitration thread throw.

Reviewed By: skyelves, xiaoxmeng

Differential Revision: D75500404

fbshipit-source-id: 2aaf145f1c1186dbf8b4d6c90a5125febe36760b