fix: Crash when aggregate push down applied on updated column with sparse row set (#13503)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13503

We was passing the effective row set instead of the input row set when
an updated column is passed to a hook.  Since the input row set is guaranteed to
be dense, the effective row set (if different from input) is larger and access
un-mapped memory then crash.
bypass-github-export-checks

Reviewed By: oerling

Differential Revision: D75466142

fbshipit-source-id: dc5d963c842bad360a23dbcf92c25b4644070d15