fix: Hold string ids for buffered inputs (#13536)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13536

Queries sometimes fail in prod with the error from StringIdMap module:
```
it != idToEntry_.end() Trying to add a reference to id 43648884 that is not in StringIdMap Split
```
It happens ~20-60 times a day, usually during high memory pressure:
https://pxl.cl/7lfgh

## Root cause
For SSD cache files and regular files we use integer IDs instead of raw filenames. Those filenames are stored in StringIdMap with reference counters. Having StringIdLease ensures that the string wll not be released.

Where do we actually hold those StringIdLeases for filenames?
1. Locally on the stack when we open the file: https://fburl.com/code/dc3q24rm
2. In `FileCacheKey` structs in AsyncDataCache module. https://fburl.com/code/lcf2u3k1
3. In `FileCacheKey` structs in SsdFile module: https://fburl.com/code/zzofy9hi
4. in LRU cache in FileHandleFactory module: https://fburl.com/code/rdyt5uw6

(2-3) records can be evicted from cache during high memory pressure: https://fburl.com/code/cbr5a3ni
(4) can be stale in LRU cache

So the only string holder we can rely on is (1) - local holder for the FileHandle on the stack.

Bus as we can see here: https://fburl.com/code/n1cm3eww
We don't pass and hold StringIdLeases to BufferedInput classes, we just copy file ids. So when we read that file, the string in StringIdLeases may be already dropped.

## Fix
The fix is simple, we just need to hold StringIdLeases while we have plans to use the file.

## Reproduction and testing
I wasn't able to reproduce it on verifier clusters, because we need to overload the memory in a cluster to trigger the cache eviction, but I was able to write a simple test.

Reviewed By: zacw7, yuandagits

Differential Revision: D75574001

fbshipit-source-id: ce8d39f3287425a5b6240355d75f35ef57704706