fix: Fix the unnest output child column size in case of identity mapping (#13584)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13584

Unnest operator produce the wrong child column vector size in case of identity map. It simply output the entire base vector in that case. This might cause problem if we only output a subset of unnest inner rows because of output batch size control. The unnest column size might not match the replicated column vector size. This PR fixes this by returning a slice of base vector in that case.

Reviewed By: wenqiwooo, Yuhta, breeze1990

Differential Revision: D75748329

fbshipit-source-id: 2886a10a46d54b3fbbfcee7cbe6177f10bc37bdf