misc: Enhance intermediate type support in presto query runner (#13462)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13462

This update improves the existing support for intermediate types, such
as "timestamp with timezone," in the Presto query runner. The
enhancement aims to simplify the process of adding support for
additional intermediate types in the future.
For types that do not require special handling during conversion
to/from their target types, this change introduces a default
implementation. This implementation leverages the CAST expression
to facilitate the conversion process.

Additionally, a check enforcing validity on offsets and sizes for null rows has been removed from the VectorMaker utility class, as these values are not evaluated when marked as null.

Reviewed By: kevinwilfong

Differential Revision: D75232290

fbshipit-source-id: 5fc01174c12c41944247abbaafd703b4fb4c8130