feat: Support single node tracing (#13612)
Summary:
The TaskTraceMetadataWriter serializes and writes the entire plan
fragment to the task metadata file. Then, the TaskTraceMetadataReader
deserializes the task metadata, extracts the target trace node, and constructs
a replay node by using it. During this process, there may be some plan nodes
that do not support serialization and deserialization in the whole plan fragment,
and it would break the tracing of the target tracing node. Hence, we need only
to trace the target node and replace its source nodes with DummySourceNode
to retain the trace node's input types.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13612

Reviewed By: gggrace14, tanjialiang

Differential Revision: D75929181

Pulled By: xiaoxmeng

fbshipit-source-id: 7a934d39890d9e5e23ca76f37de1f536cc865af8