feat: Allow access to expression stats via operator stats (#13639)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13639

The current Task abstraction has a limitation when it comes to
operators that evaluate expressions - their expression statistics
are not directly accessible. The only way to access those is through
the ExpressionListner API. This change allows access to those
expressions stats via the operator stats. Since fetching these stats
can be expensive (especially if they are fetched often) a query
option `operator_track_expression_stats` is added to allows the user
to explicitly enable them.

Reviewed By: DanielMunozT, kgpai

Differential Revision: D75997924

fbshipit-source-id: 5fe62971d5e73a0f9f46be0f87def7aaa5dc20f0