feat: Optimize unnest operator to produce one output for each input (#13687)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13687

For sequence storage, we use unnest operator to per-event type optimization which might produce a number of output batches for each input by default which is not efficient for the followup index lookup. This PR introduces a config to enforce unnest operator to produce one output for each input for index lookup efficiency.

Reviewed By: tanjialiang

Differential Revision: D76196529

fbshipit-source-id: 769810efa20dc2315b914e80cb30c17d85177f33