fix: Ensure json_extract handles invalid json during parsing (#13696)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13696

The current implementation of json_extract relies on simdjson's
document parsing API to detect invalid JSONs early in the process.
However, this API only catches some invalid conditions as per the
documentation.
To handle errors during processing, non-exception throwing APIs are
used, which return an error code instead of throwing an exception.
Unfortunately, not all calls utilize these APIs, resulting in
exceptions being thrown when trying to access invalid portions of the
JSON.
This combination of factors allowed a specific type of invalid JSON
to bypass the initial check and encounter the exception-throwing
accessor API during parsing. As a result, the UDF threw an error
instead of returning a null result as expected. This change addresses
and resolves this issue by updating those exception throwing
callsites to use the nonexcept versions instead.

Reviewed By: kgpai

Differential Revision: D76276107

fbshipit-source-id: 789fe2bf69e443186eff7bf07ef6d06774aef1de