misc(fuzzer): Add min_by/max_by to skip function list in window fuzzer when comparing result with Presto (#13729)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13729

min_by/max_by(x, y, n) function returns an array of x for the least minimal/maximal y.
The order of elements in the result array is not guaranteed when there are ties of y.
The window fuzzer reports result mismatch from Presto because of different order of
elements in the result array. So this diff blocks these two functions when window
fuzzer compares result against Presto.

Reviewed By: Yuhta

Differential Revision: D76380448

fbshipit-source-id: 51eee6dcae5cfb8b89c81bc1b8fed0d80e1e7ff0