fix: Hanging MultiFragmentTest.maxBytes (#13743)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13743

`MultiFragmentTest.maxBytes` is hanging after folly random number
generator update (due to different `Destination::targetSizePct_`).  This is because the sizes of the serialized data changed and
the test enters into a deadlock state due to the setup in the test itself (not
prod code).  Unblock this by adjusting the data sizes.

Reviewed By: bikramSingh91

Differential Revision: D76459943

fbshipit-source-id: 90557b9f99a78ebe24f97f580c5add6ff2e23157