fix: Fix RowNumber Fuzzer (#13764)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13764

RowNumber Fuzzer's source operator is a Values node which has an input that uses  an external memory pool (RowNumberFuzzerBase::pool_). This pool does not have proper reclaimer linked for driver execution, meaning when this pool triggers arbitration, it is unable to suspend the driver, making the following driver suspension check during arbitration process fail. The fix is to link the external memory pools with reclaimers having proper driver suspension mechanism, not the default memory reclaimer.

Reviewed By: xiaoxmeng

Differential Revision: D76602923

fbshipit-source-id: fb3bcee07eee24b2059ae646ba9284f30a28aa2f