feat: Reduce HashTable load factor from 0.875 to 0.7 (#13694)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13694

Load factor of 7/8 is too high. When table is 7/8 full a significant performance
degradation is observed. Highly selective use cases (when value is not
present in the hash table) are impacted the most.

According to the simulation when table is 7/8 full a single probe operation has
to visit 2.88 buckets on average.

When double hashing algorithm is used the average number of bucket to visit is
reduced to 1.88 with the idential load.

The optimal load factor seems to be at 0.7. Past that point performance starts
to deteriorate fast. Performance deteriorates much faster for linear probing
algorithm.

Reviewed By: Yuhta

Differential Revision: D76267085

fbshipit-source-id: aca39079f7cc15652b92f5979033fb77717fb188