Add fetch_size for oracle getTables to improve the performance
Oracle connector fails with a connection timeout, while show tables, when the schema has a 1 lakh tables. This happens because we are not specifying the fetch_size with the getTables method in Oracle. As a result, the connector uses the default fetch_size, which is a smaller value (10), leading to multiple network calls to the datasource and ultimately causing a connection timeout.,

Co-authored-by: lukmanulhakkeem <lukmanul.hakkeem.a@ibm.com>