[native] Enhance macOS CI build and speed it up
macOS-15 is added to the CI pipeline to match Velox builds.
In addition, the job is restructured to use more multi-threading during the
build as well as defer most dependency needed to bundling because
it has been shown that the majority of time is spent building and installing
the dependencies.