feat: Add Async SpillMerger in LocalMerge (#13634)
Summary:
The current `SpillMerger` is single-thread sync mode using `SpillMergeStream`.
It is beneficial to use multithreaded async mode to parallel the processing of
merging on the consumer side and deserialization, and IO on the producer side.
Create an async producer for each merge source using futures and callbacks, and
execute them using the folly executor of the `SpillConfig`. Unify the consumer side
by using `SourceStream` and `LocalMergeSource` in both `SourceMerger` and `SpillMerger`.

Part of https://github.com/facebookincubator/velox/issues/13260

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13634

Reviewed By: tanjialiang

Differential Revision: D76793174

Pulled By: xiaoxmeng

fbshipit-source-id: aeff30604589f894071f462a0b956346994579b1