misc: Optimize arbitrary streaming aggregation data copy (#13834)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13834

Avoid copy for arbitrary aggregation if all the sources are from the same. We might need to add dictionary if the source references are not consecutive, and otherwise we just reference a slice of source. It reduces both cpu and memory of the streaming aggregation execution more than 10x

Reviewed By: Yuhta

Differential Revision: D77003688

fbshipit-source-id: 05b0c3a64a12ca9de7f45e2a7662439317ab3453