fix: Fix semi join result mismatch with filter and multi duplicated matched rows (#13121)
Summary:
In a semi join, only the first matching record is retained when there are multiple matching rows on the other side. [ PR#13096](https://github.com/facebookincubator/velox/pull/13096) addresses this issue by selecting only the last matched row for the final output, which resolves result mismatch issue when no filter expression is present. However, this approach can not result mismatches if a filter is applied.
For example:
```
Left Record	    Right Record
a      b	       c      d
2      5	       2      4
                       2      5
```
With the join condition `a == c `and the filter `b > d`, selecting only the last matched row results in right side `(2, 5)` which does not satisfy the filter, leading to an empty result. The correct result should be `(2, 5, 2, 4)`.

This PR ensures that all matched rows are retained when a filter expression exists. During the filtering process, only the first matched row is kept.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13121

Reviewed By: kevinwilfong

Differential Revision: D77203870

Pulled By: xiaoxmeng

fbshipit-source-id: 0dcbf57148309347770aa981dec27fd148525d44