feat(cudf): Pushdown filter requirements for cuDF Presto integration (#13797)
Summary:
We are integrating velox cuDF features into Presto.

When using the cuDF parquet connector in Presto with parquet pushdown, the `outputType_` may not contain all the columns that Hive is instructed to read, and the creation of the filters may fail.

This patch ensures that any column present in the subfieldFilter or remainingFilter is read by the parquet reader and taken into account when generating the corresponding AST expression.

cc: zoltan

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13797

Reviewed By: kKPulla

Differential Revision: D77369927

Pulled By: mbasmanova

fbshipit-source-id: f7bad99538ae75fa7eb4063f6d6e2ce2677b1cd6