test(build): Add mTLS test certificates and generation script
Adds a new 'generate_mtls_certs.sh' script to automate the creation of a self-contained set of certificates (CA, server, and client) for mTLS integration tests.

This commit also includes the pre-generated certificates in the 'data/mtls_certs' directory to be used by the test suite. The README has been updated with details on the generation process.