refactor(buffer): Improving fillNewMemory tight loop (#13883)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13883

fillNewMemory is called often when Velox buffers are allocated.

We observed this method showing up on some profiles.
At the moment, it is branching inside a tight loop.
However, this branching will be the same for every iteration.

Somehow, the branch predictor is not doing its job well, as we see it profiling.
After extracting the brancing out of the loop, we almost don't see this method anymore.

Reviewed By: Yuhta, sdruzkin

Differential Revision: D77160746

fbshipit-source-id: fe5750735f7d30ae6c2ee8eb58991868d55718f7