misc: Move arrow logging include to avoid potential macro errors (#13887)
Summary:
Arrow logging defines macro DCHECK which can be redefined and result in an error if VeloxExceptions is included after it.

This PR removes the logging.h from Hashing.h because it is not used there.
It adds it to Encoding.cpp due to use of some arrow macros. VeloxExceptions
is not used and doesn’t cause a problem there.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13887

Reviewed By: zzhao0

Differential Revision: D77384117

Pulled By: Yuhta

fbshipit-source-id: e90725f9ac24b3519ee514bae365674ba5262ec2