fix: Do not crash when reading struct child larger than number of top level row (#13894)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13894

When we read file that has child vector larger than struct parent
vector (unlike in-memory case, this is suboptimal in file because child cannot be shared and the extra rows
in child would never be used), and when we apply filter on it in fallback mode
(e.g. delta update), we use the size of child vector instead of parent and
result in out of boundary access (because the result bitmask is created using top level row count).

Reviewed By: bikramSingh91

Differential Revision: D77337719

fbshipit-source-id: 01a1eb9f817e2bfa1d6ff43858a45d71c42819a9