Make rowid optional in DELETE query plan (#25284)
Summary:
Pull Request resolved: https://github.com/prestodb/presto/pull/25284

Currently, the $row_id column is automatically inserted into the output variables for DELETE queries by QueryPlanner.  If a connector does not actually use $row_id to implement DELETE, then we should not require it.  This makes $row_id optional.  If the Optional is empty, then we don't need to project the output variable.

Differential Revision: D76325048