fix: Fix the server crash caused by remote exchange error (#13905)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13905

The remote exchange might run into error so merge operator check if a remote merge source is blocked or not, it might trigger merge source to fetch data from the exchange client which in turn try to dequeue from the exchange source. The dequeue operation might throw if there is already an error. The recent merge spill support moves the merge source is blocked check inside SCOPE_EXIT which converts an throw to a server crash as we can't throw inside a dtor. This changes move the merge source is blocked check out of SCOPE_EXIT and add a unit test to reproduce and verify

Reviewed By: tanjialiang

Differential Revision: D77423282

fbshipit-source-id: b9fd6a62749a8e344b935a94ed46b74016fd4eb2