fix: Fix anti join result mismatch with filter and multi duplicated rows (#13123)
Summary:
For an anti join, if a row in the left table has multiple matched rows in the right table, and these matched rows are distributed across different batches, we need to traverse all matched rows in different batches to determine the final result.
For example, in the case below, the left table has columns a and b, and the right table has columns c and d. The join condition is a == c and b < d:

```
a b    c  d
1 3    1  2
       1  2
       1  4
```
Meanwhile, the output batch size is 2, so the first batch1 is:
```
a b c d
1 3 1 2
1 3 1 2
```
If we do not wait for the subsequent batch, such as `1 3 1 4`, the result for batch1 would output a record `1 3`. However, the final result should be null because `1 3 1 4` satisfies `b < d`.

The current PR relies on the joinTracker mechanism to ensure that the final result is output only after all matched rows have been traversed.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13123

Reviewed By: kevinwilfong, Yuhta

Differential Revision: D77379609

Pulled By: xiaoxmeng

fbshipit-source-id: b9475c0bb0d84baf3ac0b3e5c9245553c9996a1f