feat: Add prefix key test in IndexLookupJoin to make sure it supports prefix keys (#13902)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13902

Add the prefix key test in IndexLookupJoinTest as the first step to verify that IndexLookupJoin Operator is opache to support prefix key indexLookupJoin.

The core logic of handling should be determined by the concrete connector implementation, in this case, the SequenceStorageConnector, and further SequenceStorageIndexDataSource and client library to coomunicate with zippy.

Next step is to write test in SequenceStorageIndexDataSourceTest to find out logic where need to be changed and further more is the library communicate with zippy.

Reviewed By: xiaoxmeng

Differential Revision: D77414916

fbshipit-source-id: 72877723e96a7b5a14352ca5f9b17c8629935365