feat: Add null key support for index join (#13891)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13891

Add null key support for index join. If a probe input has any null in the columns used for index join, the index join operator will skip the join and return the null result.
Also fix an invalid check in partition serde which expects the child vector has the same length as the parent row vector which is not always the case for velox vector. We shall only require the child vector is no less than parent vector size

Reviewed By: breeze1990, zacw7

Differential Revision: D77202778

fbshipit-source-id: 3f762f3ebdd84d3b1ffdadce4e58e3a78d77b0e3