feat(noisy_sum): Add support for all numeric types in noisy_sum_gaussian(col, noise_scale) (#13702)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13702

## This Diff

This diff adds support for all numeric types in the `noisy_sum_gaussian` function.

### Key changes

**Testing**
New test cases have been added to cover the additional numeric types.

**Implementation**
The implementation of the `noisy_sum_gaussian` function has been updated to handle the new numeric types. This includes dispatching to the corresponding type for the update sum operation.

**Impact**
This diff extends the functionality of the `noisy_sum_gaussian` function to support more numeric types, making it more versatile and useful in a wider range of applications.

Reviewed By: duykienvp

Differential Revision: D75964171

fbshipit-source-id: a5aaca6bdc4dc29702369528767f344c88f1c24c