feat(noisy_avg): Add support for optional lower and upper bounds (#13710)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13710

The new function `noisy_avg_gaussian` allows users to specify lower and upper bounds for the input values.

**Overview**

This diff introduces a new feature to the `noisy_avg_gaussian` function. The function now supports optional lower and upper bounds for the input values before averaging.

The bounds are applied before averaging, and the average is then calculated using the clipped values. The average is then perturbed with a normally distributed random double value with 0 mean and standard deviation of `noise_scale`.

**Code**

The changes are made in the following files:

*   `NoisyAvgGaussianAggregationTest.cpp`: Test for the new feature.
*   `NoisyAvgAccumulator.h`: Add support for lower and upper bounds.
*   `NoisyAvgGaussianAggregate.cpp`: Modify the function to handle lower and upper bounds.
*   `aggregate.rst`: Add documentation for the new function signature.

Reviewed By: duykienvp

Differential Revision: D76209271

fbshipit-source-id: 381c307184bfcbdce69ede329e34a21d9243fb9b