fix: Do not use hook in DeltaUpdateColumnLoader (#13996)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13996

In `DeltaUpdateColumnLoader`, hooks cannot be used efficiently and only
used as a fallback.  Since hooks are only used in aggregation today, we can add
a new method `VectorLoader::supportsHook` to indicate if hooks should be used,
and avoid using of hook in aggregation if the loader does not support it.  This
allows us to remove some inefficient and error-prone code (some indices mismatch
is causing out of boundary access and crash).
bypass-github-export-checks

Reviewed By: oerling

Differential Revision: D77618589

fbshipit-source-id: 71bef3cc8bef1988b236d336bc9f92d09a020c91