fix(vector)!: Ignore size parameter for FlatVector::mutableValues (#14012)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14012

Today FlatVector::mutableValues takes a size parameter and ensures the values Buffer in
the FlatVector is sized to hold at least that many values.  If the Buffer is null, shared, or is a view, a new Buffer of that size is allocated.

This can result in the FlatVector ending up in an invalid state, if the size passed in is smaller
than the size of the Vector and a new Buffer is allocated.  In this case, if we call FlatVector::set, FlatVector::valueAt, or other functions that access the value at an index, this
can read or write off the end of the Buffer even for an index that seems valid based on the
size of the Vector.  All of our checks within FlatVector (or at least all that I saw in a cursory
inspection) are based on the size of the Vector, not the size of the values Buffer.

This can be fixed by silently using the maximum of the size passed in or the size of the
Vector. This is what BaseVector's mutableNulls does for example. I don't see any need for
passing in a size greater than the size of the Vector, in particular I don't see any use cases
for this in the existing calls to  mutableValues.

Therefore I've opted to simply remove the argument.

Since Nimble depends on this API call, I have to make this change in 3 steps:
1) This change to make the argument optional and ignored.
2) Update Nimble to pull in Velox with this change and remove the argument from the call to
mutableValues
3) Remove the argument entirely from mutableValues in Velox

BREAKING CHANGE:
This change ignores the size parameter from FlatVector::mutableValues.  Callers will need to
remove the argument from the function call and ensure they've sized the Vector
appropriately before calling mutableValues.

Reviewed By: Yuhta

Differential Revision: D77766602

fbshipit-source-id: b5e1f25232bc17cc7ffbd6605c0d5af28096462f