misc: Add ability to track stats on non-special form expressions (#14037)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14037

Recently a query config `operator_track_expression_stats` was added to expose
expression stats via operator stats, however, the stats exposed turned out to
included every expression (constants and field references, etc) and became
unwieldy. For the use-case that motivated this change, we are only interested
in expressions that evaluate a UDF and therefore encompass all non-special
form expressions. We therefore update the query config to only track these
kind of expression. In the future, if we decide to expose more options, we can
enhance this config to support those.

Reviewed By: mbasmanova

Differential Revision: D77903975

fbshipit-source-id: 2ae24c38e5ae824bcf27a10baab3f0807bd86e53