[pos][native] Specify custom airlift version for pos modules
Presto is moving to JDK17. As a part of this we are also upgrading airlift
dependency to 0.219. But 0.219 is not backward compatible with older JDK 8/11.

Presto-on-Spark (atleast internally at Meta) runs on Spark2.4(JDK8) and Spark3.4(JDK11)

Hence packaging the 0.219 version of airlift in presto-on-spark builds is going
to result in runtime incompatibility of  airlift and the JDK vesion.

To mitigate this issue, i.e. to let presto-on-spark builds package a custom version
of airlift, we override the airlift version in presto-spark-package pom.

Currently we stick to 0.216 as it is known working version for JDK8.