refactor: Move string-to-decimal API to DecimalUtil (#13253)
Summary:
Move the `toDecimalValue` API into DecimalUtil so it can be reused in the Spark
from_json function and the decimal partition key support.

The below fixes are added for the string-to-decimal conversion:
Returns error status when the exponent part is empty.
`Status::UserError` internally uses `fmt::vformat` which throws  below
exception when the string contains placeholder `{` or `}`. This PR removes the
details from error message.
   > uncaught exception of type fmt::v10::format_error: unmatched '}' in format string

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13253

Reviewed By: Yuhta

Differential Revision: D77948132

Pulled By: kagamiori

fbshipit-source-id: a12bb36824565a1eebb2e27b0d7c85a5f784e9ff