fix: Cast to JSON has wrong output size estimation for some invalid Unicode input (#14062)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14062

When the input string is not valid Unicode, the size estimation skips all 4 bytes but we should only skip 1 byte and re-parse the input from second byte.  The actual conversion is done correctly, but the size estimation is wrong due to this mismatch.  This is causing out of boundary write and crashes.

Reviewed By: xiaoxmeng

Differential Revision: D78028993

fbshipit-source-id: b5d51abf6c8ef8dadc7a0eed26587227a76c3309