feat: Implement utility library FilterToExpression (#13936)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13936

Implement utility method FilterToExpression which will be used to convert Expression from corresponding Filter

Because tableEvolutionFuzzer is using generateSubfieldFilters which returns SubfieldFilters, thus, in order to be used in later in the test, we need to convert Filter to TypedExpression instead of strings to TypedExpression. The reason to do it via a utility file is that this maybe used by other similar scenarios in the future.

Reviewed By: Yuhta

Differential Revision: D77549648

fbshipit-source-id: 175ffe068fb128719dfcac1632b543442492e641