feat: Extend unnest to support left join semantics by producing output for input row which has empty/null unnest value (#14095)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14095

Add option in Unnest plan node to support yield one output row for the input row with all empty or null unnest values. If configured, the output has one additional column to indicate if the output row has empty unnest value or not. If it is empty, then the corresponding unnest columns are all set to null for that row. extractRowRange is extended to ensure the buffer allocations for  the output rows with empty unnest values. This is to extend to support left join semantics for unnest operation which only supports inner join semantics. The left join semantics is required for Meta internal AI index join use case and Spark gluten use case as well: https://github.com/facebookincubator/velox/issues/9630.

Reviewed By: mbasmanova

Differential Revision: D78238081

fbshipit-source-id: 3b60f5dff73788f67636d251768e0507212e31e2