feat: Add TokenProvider to ConnectorQueryCtx and include it in the cache key for FileHandle (#13919)
Summary:
Summary
As discussed in https://github.com/facebookincubator/velox/issues/13875 and https://github.com/facebookincubator/velox/issues/13914, we need to provide user information to the filesystem to support authentication and impersonation. This requirement can be generalized with a TokenProvider abstraction.

This PR introduces an abstract `TokenProvide`r class and integrates it into both `ConnectorQueryCtx` and `FileOptions`. It also becomes part of the `FileHandle` cache key.

To use it, users simply need to implement the TokenProvider interface with hash, equals, and getToken methods. The filesystem can then access relevant user or token information when reading from or writing to files.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13919

Reviewed By: Yuhta

Differential Revision: D78280984

Pulled By: xiaoxmeng

fbshipit-source-id: fa26c566aef45a66863e0328ad633b27a974f218