refactor: Use size_t/intXX_t instead of int (#14114)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14114

Use the semantically more correct size_t instead of int for vector
indices.  Also replace int with an appropriate intXX_t for enums/etc.  This reduces
the platform dependences and also removes lots of lints.

Also reserve some vectors that we create in a loop.

The first version of this caused segfaults because the switch from int to size_t
increased the number of bytes written by VarbinaryWriter for a value from 4 to 8,
which messed up serialization.  Static casting size_t to int32_t fixes that.

Depends on D78418978

Reviewed By: bikramSingh91, kgpai

Differential Revision: D78415824

fbshipit-source-id: 3e567022116bd7a5736ad544b257cef9bb371851