velox(feat): unblock Tpch for native query execution (#14187)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14187

as part of the Axiom project, we are adding an end-to-end query execution pipeline in C++. There are two changes which need to be made to make this possible with a Tpch backend:

1. need to be able to pushdown filters into the TPC-H data source
2. need the ability to distinguish between TPC-H tables of different scale factors using the public ConnectorTableHandle APIs

For (1), we can store the filters in an optional field inside the TableHandle and pass them down through the data source. Then, when reading from the data source, apply filters to the rows generated via the Tpch generator.

Rather than deal with the complexity of subfields, I am combining all filters in a `remainingFilter` expression and evaluating that directly. If a filter was not emplaced into the parent table handle, the data source path remains the sam

For (2), modifying the `name()` API to return a fully-qualified name inferred from the scale factor rather than just the name of the base table. Anything with a scale factor <1.0 is resolved to the `tiny` table

Reviewed By: yuandagits

Differential Revision: D78605275

fbshipit-source-id: 599ab7bb1eab6095a6c6f5ba8812c03a11bb71ec