fix: ConstantTypedExpr::equals null behavior (#14191)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14191

Fixes ConstantTypedExpr::equals for constant expressions with null values. Currently, we simply return the value of BaseVector::compareNulls, which operators like a comparator (0 if same 1/-1 otherwise), which returns the opposite value we want (1 or true if same, 0 otherwise).

This will also solve the issues and reduce the flakiness we see in fuzzers because of this error: https://github.com/facebookincubator/velox/issues/14122

Reviewed By: kagamiori

Differential Revision: D78771083

fbshipit-source-id: 3fff4fbd54727fe78161dd7371badf0954f50a39