Filter signatures from only list if they are also part of skiplist (#14196)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14196

Currently the bias aggregation fuzzers fail if the --only functions list has functions that are in the skip list. For e.g : https://github.com/facebookincubator/velox/actions/runs/16427655994/job/46426205042?pr=14117 . We fix the Aggregate fuzzer to filter signatures that are also part of the skip list.
This change filters out signatures from only list if they are present in the skip list.

Reviewed By: peterenescu, oliver6782

Differential Revision: D78793063

fbshipit-source-id: 540e6b8607dcd8c5f4a9f399b99397fab5bf09d4