feat: Add output pool parameter in LocalRunner constructor (#14225)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14225

In LocalRunner, we lack the ability to pass in param_.outputpool. As a
result, we create a new output pool for each LocalRunner, additionally,
the pool would be freed during LocalRunner desctruction.

However there are cases where we want to store the output vector
which means the output pool associated with the output vector need
to have longer lifespan than LocalRunner itself.

Add the ability to pass in output pool in LocalRunner constructor

Reviewed By: oerling

Differential Revision: D78873789

fbshipit-source-id: afc0c8ccf807db68dfa2167c395440c062b5f8f6