misc: Improve Barrierred execution error string (#14233)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14233

Currently when barriered execution failed due to unsupported PlanNode, the only error string is "Task doesn't support barrier". It is hard to debug which node doesn't support it. This diff makes it easier by including first node that doesn't support barriered execution in the plan fragment in the error message.

Reviewed By: xiaoxmeng

Differential Revision: D78950605

fbshipit-source-id: 368f93863c64af74549aaf6c3b368a1de8c036d7