misc: Enable decimal expression fuzzer test in workflow (#14044)
Summary:
The decimal expression fuzzer test had been enabled in the experiment workflow,
which was removed with https://github.com/facebookincubator/velox/commit/331b82de2b423af464ced0422df61d7009d6cce8. This PR enables it in the scheduled workflow.

Fixes the fuzzer failure issue for Presto `parse_presto_data_size` function.
When `parse_presto_data_size` is nested inside another function `A`,the return
type of `parse_presto_data_size` is decided when fuzzing argument types for `A`,
which is random and causes a mismatch with the expected DECIMAL(38, 0).
This PR fixes this issue by respecting integer constraints in the
`checkOrSetIntegerParameter` so as to skip the test for this function.

Includes ParsePrestoDataSizeTest.cpp in the CMakeLists.

Fixes https://github.com/facebookincubator/velox/issues/13144.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14044

Reviewed By: peterenescu

Differential Revision: D79044881

Pulled By: kagamiori

fbshipit-source-id: 5eb04ab6d565a236cf495aa08099007365cd61ba