feat: Support comparing MapVectors and FlatMapVectors (#13940)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13940

Given that MapVectors and FlatMapVectors have the same Type, we may find ourselves
needing to compare maps in Vectors of each encoding. E.g. this comes up frequently in
testing.

This diff adds support for this to FlatMapVector's compare function. For the sake of
consolidation MapVector simply flips the order and invokes FlatMapVector's compare
function.

Reviewed By: pedroerp

Differential Revision: D77563385

fbshipit-source-id: 6677fc3293759738f3d16b6368d22dd6c0bd9c1c