fix: Do not keep reference to input vector in FilterProject longer than needed (#14269)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14269

`FilterProject` always consumes all the input by the time we return
from `getOutput`, so there is no need to track number of rows consumed.  In the
current implementation, this was causing us to hold on to the input after
`getOutput` is returned and preventing memory reusing when the batch is being processed by downstream operators.

Reviewed By: mbasmanova

Differential Revision: D79174768

fbshipit-source-id: aa4a41121df80cb59ad53bfdab1f876c9c72c1b8