fix: Respect Greek final sigma rule in Spark lower function (#13993)
Summary:
Respects the Greek final sigma rule in Spark lower function. For the uppercase
letter Σ, if it appears at the end of a word, it becomes ς. In all other
positions, it becomes σ. If false, it is always converted to σ.

Renames the template parameter of lowerUnicode as `turkishCasing` and adds it
in the `initcap`.

Fixes https://github.com/facebookincubator/velox/issues/14155.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13993

Reviewed By: kgpai

Differential Revision: D79266088

Pulled By: kKPulla

fbshipit-source-id: 43720129f72fca6df14a6d8c990a09c924ddc75f