feat(cudf): Special handling of VARCHAR literals in expression trees (#13909)
Summary:
For a special case where the expression is a single string literal, `cudf::compute_column` fails to create a column out of it because it cannot output variable width types. This PR adds special handling for this case and uses a cudf column factory to stand in.

This fix is required for TPCH query 21 to work in presto.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13909

Reviewed By: kKPulla

Differential Revision: D79371272

Pulled By: kgpai

fbshipit-source-id: 3dc1fb6b3c7a203bbc2eb4eafa76663519de3e1d