feat: Add match column for the left index lookup join (#14373)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14373

Support an optional match column for index lookup join operator to indicate if an output row has a match or not. If this option is set in the plan node, then we expect a boolean column at the end of the index join output, otherwise not. This only applies for left join type.

Reviewed By: zacw7, mbasmanova

Differential Revision: D79839444

fbshipit-source-id: 9fa804a3beb12aa7b411aaaff2e8afeb474d17d6