fix: SubscriptUtil doesn't respect map offset when caching complex keys (#14426)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14426

In the Velox UDFs for map subscript and element_at we use the SubscriptUtil library for the
implementation.

In the implementation in SubscriptUtil when we enable caching and the key is a ComplexType
we create a HashMap where the key is a structure that contains the key's base Vector and an
index into it. Currently we assume that index starts from 0, but it should really be based on
the offsets in the MapVector.

This change corrects this.

This addresses https://github.com/facebookincubator/velox/issues/14395

Reviewed By: kagamiori

Differential Revision: D80021449

fbshipit-source-id: cb64cf66440da1982e2e1ef16562b1ea9ff6e995