feat(expr): Support non-ANSI Spark semantics for cast failures in complex types (#14451)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14451

In Spark with ANSI mode turned off, when casting the elements of a complex type from one
type to another, if the cast fails, that particularly element is null in the result.

This differs from the current semantics of Velox's CastExpr (with try cast enabled) which
follows the ANSI standard, which is to return null as the result for the entire complex type.

E.g.
TRY_CAST(ARRAY('1', 'abc', '2') as ARRAY(BIGINT))

This currently returns NULL in Velox. In Spark this would return ARRAY(1, NULL, 2).  It's also
worth noting Spark would return this whether it was a TRY_CAST or a CAST.

This change adds a setting in the CastHooks `applyTryCastRecursively` which when true
applies the Spark semantics where a failed cast only effects the value in a complex type that
caused the failure.  This is false in PrestoCastHooks and true in SparkCastHooks.

Reviewed By: kgpai

Differential Revision: D80135098

fbshipit-source-id: ac3eafd26f3414f6315e5739e3410a6fcfb1d29b