Cancel duplicate checks release notes actions
The action to check release notes is triggered by each edit
to the description. For example, selecting the check boxes counts
each as a single change. It triggers an action for each selection.
Thus, up to 6 actions to check the release notes might be queued.
This change cancels other queued actions for the same PR.