fix: print bucket and key in s3 object already exists error message (#14347)
Summary:
When I used the TableWrite operator of velox, a logic error caused the same file to be uploaded to the object storage multiple times. S3WriteFile discovered this error, but I found that I could not tell from the log which object was uploaded multiple times. This PR enhances the error printing to deal with this scenario.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14347

Reviewed By: kKPulla

Differential Revision: D80122164

Pulled By: pedroerp

fbshipit-source-id: e8997c088735a6fb56b15c081c17dc208f291340