feat: Add variant of project node that splits subfields of struct without materialization (#14461)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14461

In order to fully utilize `ParallelProject`, we need to keep the
subfields lazy when we lift them to top level columns, so that we can
parallelize the decoding.  This was not possible because the normal evaluation
framework would load the whole column vector tree at beginning of the
evaluation.

We address this by providing a variant of project node called `LazyDereferenceNode`.
This node contains only field access projections, and the output of
them should have no overlap.  This way all the outputs can be kept lazy
whenever possible.
bypass-github-export-checks

Reviewed By: oerling

Differential Revision: D80176959

fbshipit-source-id: e29900bfe23ea9edd14cbf7d9bcc4547dcf9fe4c