feat(Velox): Expose Filesystem Stats as Runtime Stats in Hive Data Sink (#14465)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14465

In HiveDataSink, expose the Filesystem Stats (IoStats) as runtime stats. This matches the existing logic for HiveDataSource.

This prepares for these metrics to be exposed as TableWriter operator stats

## Context

Filesystem Stats (IOStats) is generic stats tracker which filesystems may add arbitrary stats to. This is intended to support tracking IO statistics which are specific to the filesystem implementation (as opposed to the generic IoStatistics, which are included in open-source Velox).

This was previously only supported for Read operations. These PRs add support for Write operations as well.

Reviewed By: Yuhta

Differential Revision: D80153413

fbshipit-source-id: 4e370d680308d4887df99cdd8cc75b5083731de4