Parse CharN in protocol to Varchar type
Reuse the existing Velox VarcharType to implement the type Char(n) in protocol.
Add a SystemConfig "char-n-type-enabled" to guard this feature.
Note this will make Char(n) type carry the behavior of VarcharType type. It is a different
behavior from Char(n) type in Presto today, where it has a fixed number
of characters. We suppose the user could call rpad() if today's behavior is needed.