refactor: Make local merge source queue size configurable (#14540)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14540

Add the new query config `local_merge_source_queue_size` to control how
many vectors we buffer in a local merge source.  This can be used to control the
memory usage of the pipeline.

This change does not alter any behavior of existing applications, which continue to use the default value (2) same as before.

Reviewed By: xiaoxmeng

Differential Revision: D80626475

fbshipit-source-id: acdeeacf4267775b01412ffe3768341152e617cd