feat(wave): Improve workload balance across thread blocks in Nimble reader (#14448)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14448

This diff introduces two different dispatch modes for decoding Nimble files. When the `decodeStep`s are independent, we can dispatch them to different thread blocks in order to maximize the parallelism. On ther other hand, when they are dependent we make sure that  inter-dependent steps are dispatched to the same thread block in order to avoid dependencies between thread blocks. Two `decodeStep`s are dependent if they evaluate filters and the sets of rows they are assigned overlap. The dependency is mandatory because we evaluate all filters in one kernel by letting the filter results cascade from one column to another. Therefore, we must guarantee that the same set of rows is always processed by the same thread block.

Reviewed By: Yuhta

Differential Revision: D80126357

fbshipit-source-id: f03c27ddba7037baf8d83adbed96b0d6a08431a8