fix(udf): Update TimeZoneNames.cpp (#14560)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14560

I originally created TimeZoneNames.cpp running with OpenJDK 1.8.0_362.  This change
regenerates it using OpenJDK 17.0.2

This introduces names for more modern time zones that must have been missing in the older
version.  E.g. Europe/Astrakhan was introduced in 2016 and was not present in the older file.

I chose OpenJDK 17.0.2 as Presto was recently upgraded to requiring 17 as the minimum
version of Java
https://github.com/prestodb/presto/pull/24866

Reviewed By: peterenescu

Differential Revision: D80730213

fbshipit-source-id: 5ea19e6101416a1b27c816c81e5539419a526bc2