feat(type): Add BigintEnum type in Velox (#14182)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14182

Adding BigintEnum type as a parametric type in Velox. This type will get registered once with name "BigintEnum", but the BIGINT_ENUM type takes in a std::vector<TypeParameters> which contain the enum name and values map. This way, each different enum type can be instantiated separate from the other enums with a different name and map.

Instances are stored in a static folly EvictingCacheMap with a set max size of 1000 which matches the Java implementation.

New parsing rules will be added in the following diff to recognize the string pattern that Java coordinator serializes the type info to, and to parse that into a string enumName and a std::map<std::string, int64_t> valuesMap.

Reviewed By: mbasmanova, kgpai

Differential Revision: D78720602

fbshipit-source-id: a6a32949e0e9d83abd0e13bce91e8dd01867431e