feat: Add SpatialJoinNode to presto_protocol
To send SpatialJoinNodes to Velox, we need to serialize and deserialize them
via presto_protocol.

This change requires https://github.com/facebookincubator/velox/pull/14339 for
spatial joins to not cause an error.  After this PR and the above lands,
Spatial Joins should be enabled implemented as Nested Loop Joins.  Not
efficient, but it should be correct.