fix: Optimize projectChildren (#14594)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14594

Currently we are using a hash map to keep track of wrapped children in
`projectChildren`.  This could be inefficient in case there are many top level
columns (>1k) and showing up to 10% in CPU profile.  Since the max input channel
is not exceeding the number of top level columns, we can use a vector to
efficiently replace it.

Reviewed By: bikramSingh91

Differential Revision: D80886554

fbshipit-source-id: fd5d723709e95318ef50eeb7b8eaf05149961a01