build: Use gersemi instead of the no longer maintained cmake-format (#14237)
Summary:
Closes https://github.com/facebookincubator/velox/issues/14209

While the style deviates a bit from what we have now there are only a few options as gersemi is "opinionated". But there are no alternatives, so... :shrug:

On the flipside the extension and definition system makes it quite easy to allow our custom functions to be properly formatted as well!

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14237

Reviewed By: kagamiori

Differential Revision: D81090938

Pulled By: kgpai

fbshipit-source-id: a4cdfb9d6b19575e3586f93a216c7eab0490d92c