refactor(abfs): Separate client creations for different authentications and add AzureClientProvider interface (#14281)
Summary:
1. Currently the read/write clients for abfs are created from the `AbfsConfig`.
This patch introduces the `AzureClientProvider` interface for creating
read/write client, and the part of logic in `AbfsConfig` to handle the auth
configurations are extracted out.

2. Rename AbfsConfig to AbfsPath, which only preserves the logic that
parses the abfs file path.

3. Separate the client creations for different auth types.

4. Add `AzureClientProviderFactories` to support registering the
creator functions of `AzureClientProvider` for different accounts.

Part of the changes in this patch are extracted from https://github.com/facebookincubator/velox/pull/14045

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14281

Reviewed By: xiaoxmeng

Differential Revision: D81333358

Pulled By: peterenescu

fbshipit-source-id: 2e5b7194109f58e1a2092443be0f52e08818dbd2