refactor: TypeParser to use velox/.../prestosql/...TypeParser (#25877)
Summary: https://github.com/facebookincubator/velox/pull/14208 moves
velox/type/parser -> velox/functions/prestosql/types/parser. This is
because this type parser was intended to be used for presto specific
purposes, but was placed in a folder which seems too general. With this
merged, now moving references to Velox's type parser to the new folder.

Differential Revision: D80919225


```
== NO RELEASE NOTE ==
```