fix: Fix the partition scale and stats report condition and the scale writer test flakiness (#14707)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14707

Fix the scale partition condition to skip balancing partition which has zero bytes in case that the scale threshold bytes is also zero. Also fix two extreme cases that
(1) all written data go to very few partition so that once rebalance gets triggered, then it will trigger rebalance. The failed test case is to verify the rebalance never triggers if the memory threshold is set to low enough, fix the flakiness by reducing the ratio from 0.1 to 0.001 and verified with 2000 runs;
(2) there is time race in rebalance stats reporting across multiple drivers. The current code reports the stats when there is no other shared reference (the last closed driver operator). But this is not reliable as it is possible (from the test logs), all the drivers close at the same time and no one report the stats. Fix the production code to report the stats from the first driver operator instance (driver id = 0), this ensures we have sth to report and it is sufficient good in production. This still can cause the test flakiness if the rebalance happens while some driver starts closing. We can totally prevent such as so relax the check a bit in test for the test case that we expect rebalance to trigger. Verified with 2000 runs.

Reviewed By: mbasmanova

Differential Revision: D81610234

fbshipit-source-id: 265cc2cc3e74d3fb7798c35c6821ecd55437a0c7