fix: Add check to isInMap before casting as mutable buffer (#14612)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14612

Add check to flatMapVector function isInMap to confirm inMaps[channel] is not null before casting as mutable buffer.

inMap is allowed to be null and is encouraged for short circuiting for keys that are present in every row; however, we must confirm that it is not null before casting as mutable buffer and attempting to reference individual bits.

Reviewed By: pedroerp

Differential Revision: D81077301

fbshipit-source-id: a85c859bea5e33a4e782fa75e3e47e3271e024dd