fix: Optimize SMJ to avoid generating too many small batches from right side (#14699)
Summary:
In Meta internal AI use case, we see large performance regression which is caused by the small batches generated SMJ because the right side table has produced way more small batches than the left side. And during the output processing like search for match, we produce an small output if the current right batch has finished. This PR fixes this by materialize the right side in case that like we do for match output processing. The performance regression is fixed by this.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14699

Reviewed By: Yuhta

Differential Revision: D81562783

fbshipit-source-id: e6616523232a07a42ce63abefaf1d16c5e32b6da