[pos][native] Setup cpp worker memory settings based on sparkConf (for smart retries) In this PR we make the change where we base our cpp worker memory calculations on the config that is updated to higher values by spark runtime for retries.
Currently this config is `spark.memory.offHeap.size`, as internally at Meta,
this property is updated to higher values on retries.

Note: We could possibly consider parameterizing this so that other use-cases can benefit,
but as of today only Meta uses the presto-on-spark native codepath. So we leave this enhancement for future.

Also note that even for first attempt we use same logic, as it will just
spit out existing configured values.

This simplifies the approach, compared to the presto-on-spark/java smart retries
which only calculate memory settings based on attemptNumber to ensure logic is used
only for retries