fix: Fix the column mismatch output handling for left index join (#14772)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14772

Index lookup join doesn't fill the match column properly. It set the fill end bit to the number of bits to fill which is wrong. This is discovered by the extension to Meta internal use case. The existing index join unit test can't catch this because (1) the test always generate the hit probe rows first; (2) match verify logic only check lookup value is null if match value is false so this can't catch the issue. This PR fix the issues (1) randomize the probe hit rows; (2) check lookup value for both value null and not null cases. Verified that with this change, the improved index join unit test can catch the issue.

Reviewed By: zacw7, mbasmanova

Differential Revision: D80915369

fbshipit-source-id: f3788694542c5d66777ba40c04a492eb544c5432