fix(build): Replace actor with owner for determining merge base (#14706)
Summary:
Previously the actor (the workflow initiator) was used to try and determine the merge base commit sha. However, a user pushing to a different repo becomes the owner and $OWNER:$HEAD_REF does not exist in that scenario. Instead, try to use the PR creator name. We’ve seen HTTP 404 from GH CLI when the actor does not match the repo name.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14706

Reviewed By: mbasmanova

Differential Revision: D81922077

Pulled By: Yuhta

fbshipit-source-id: d2c81f38e0b37b62288a2b797b8954950a7e9fd0