build(ci): Fix build report (#14762)
Summary:
I thought there was an issue with the report code breaking the build but it was actually our conbench instance. It seems to have issues with fetching Github metadata (I have already started investigating).

This is a workaround but should not cause any issues as we don't collect metrics in PRs.

Fixes https://github.com/facebookincubator/velox/issues/14720

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14762

Reviewed By: mbasmanova

Differential Revision: D81921562

Pulled By: Yuhta

fbshipit-source-id: b5426134b3f166f44c904ab8704cea7ac918258d