feat(cudf): Support enable cudf in task level (#14216)
Summary:
Register cudf in executor, and may disable cudf in task plan level. For several plans, if the plan cannot fully be executed in GPU, we will not offload this stage to GPU to avoid format conversion cost, so we need this config to disable CUDF driver adapter.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14216

Reviewed By: Yuhta

Differential Revision: D82322695

Pulled By: kagamiori

fbshipit-source-id: 110e209a38a072ec6170e80fa26b804205ce7c42